// src/generated/models/Cr871_settingsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr871_settingsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr871_settingsstatecode = keyof typeof Cr871_settingsstatecode;
export const Cr871_settingsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr871_settingsstatuscode = keyof typeof Cr871_settingsstatuscode;

export interface Cr871_settingsBase {
  cr871_description?: string;
  cr871_setting1?: string;
  cr871_settingid: string;
  cr871_settingvalue?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr871_settingsstatecode;
  statuscode?: Cr871_settingsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr871_settings extends Cr871_settingsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}